base: Bring back certified props overlay
After dropping certified props overlay, PI doesn't pass out of the box. Because the
function `setCertifiedProps()` listens for overlay `config_certifiedBuildProperties`
which is set via this overlay. So, now, it should pass PI out of the box and at the
same time json based spoofing is preserved for guarding PI fail if google bans fp
in between releases.

Change-Id: I7518b943e05265a5d73a3b0bfe99b7e449efa345